Rename to monorail subcommand and start implementing with Hiffy
