+++ Cargo.toml
@@ -28,6 +28,7 @@ members = [
     "cmd/lpc55gpio",
     "cmd/manifest",
     "cmd/map",
+    "cmd/monorail",
     "cmd/openocd",
     "cmd/pmbus",
     "cmd/probe",
@@ -39,8 +40,8 @@ members = [
     "cmd/rendmp",
     "cmd/ringbuf",
     "cmd/sensors",
-    "cmd/spd",
     "cmd/spctrl",
+    "cmd/spd",
     "cmd/spi",
     "cmd/stackmargin",
     "cmd/stmsecure",
@@ -48,7 +49,6 @@ members = [
     "cmd/test",
     "cmd/trace",
     "cmd/validate",
-    "cmd/vsc7448",
     "xtask",
 ]
 
@@ -88,6 +88,7 @@ cmd-jefe = { path = "./cmd/jefe", package = "humility-cmd-jefe" }
 cmd-lpc55gpio = { path = "./cmd/lpc55gpio", package = "humility-cmd-lpc55gpio" }
 cmd-manifest = { path = "./cmd/manifest", package = "humility-cmd-manifest" }
 cmd-map = { path = "./cmd/map", package = "humility-cmd-map" }
+cmd-monorail = { path = "./cmd/monorail", package = "humility-cmd-monorail" }
 cmd-openocd = { path = "./cmd/openocd", package = "humility-cmd-openocd" }
 cmd-pmbus = { path = "./cmd/pmbus", package = "humility-cmd-pmbus" }
 cmd-probe = { path = "./cmd/probe", package = "humility-cmd-probe" }
@@ -99,15 +100,14 @@ cmd-rencm = { path = "./cmd/rencm", package = "humility-cmd-rencm" }
 cmd-rendmp = { path = "./cmd/rendmp", package = "humility-cmd-rendmp" }
 cmd-ringbuf = { path = "./cmd/ringbuf", package = "humility-cmd-ringbuf" }
 cmd-sensors = { path = "./cmd/sensors", package = "humility-cmd-sensors" }
-cmd-spd = { path = "./cmd/spd", package = "humility-cmd-spd" }
 cmd-spctrl = { path = "./cmd/spctrl", package = "humility-cmd-spctrl" }
+cmd-spd = { path = "./cmd/spd", package = "humility-cmd-spd" }
 cmd-spi = { path = "./cmd/spi", package = "humility-cmd-spi" }
 cmd-stackmargin = { path = "./cmd/stackmargin", package = "humility-cmd-stackmargin" }
 cmd-stmsecure = { path = "./cmd/stmsecure", package = "humility-cmd-stmsecure" }
 cmd-tasks = { path = "./cmd/tasks", package = "humility-cmd-tasks" }
 cmd-test = { path = "./cmd/test", package = "humility-cmd-test" }
 cmd-trace = { path = "./cmd/trace", package = "humility-cmd-trace" }
-cmd-vsc7448 = { path = "./cmd/vsc7448", package = "humility-cmd-vsc7448" }
 cmd-validate = { path = "./cmd/validate", package = "humility-cmd-validate" }
 
 fallible-iterator = "0.2.0"
