+++ cmd/monorail/Cargo.toml
@@ -1,17 +1,18 @@
 [package]
-name = "humility-cmd-vsc7448"
+name = "humility-cmd-monorail"
 version = "0.1.0"
 edition = "2021"
-description = "VSC7448 operations"
+description = "Management network control and debugging"
 
 [dependencies]
 anyhow = { version = "1.0.44", features = ["backtrace"] }
 hif = { git = "https://github.com/oxidecomputer/hif" }
 humility = { path = "../../humility-core", package = "humility-core" }
 humility-cmd = { path = "../../humility-cmd" }
-humility-cmd-spi = { path = "../spi" }
+humility-cmd-hiffy = { path = "../hiffy" }
 log = {version = "0.4.8", features = ["std"]}
 parse_int = "0.4.0"
 clap = { version = "3.0.12", features = ["derive", "env"] }
 vsc7448-info = { git = "https://github.com/oxidecomputer/vsc7448.git" }
 vsc7448-types = { git = "https://github.com/oxidecomputer/vsc7448.git" }
+idol = {git = "https://github.com/oxidecomputer/idolatry.git"}
